src/requirements.txt
datasets
click
google-cloud-translate
openai